Passage de getNumberOfComputer en jpa criteria

+++ CDB/console/src/main/java/com/excilys/cdb/console/controller/CliController.java
@@ -165,7 +165,7 @@ public class CliController {
 				.collect(Collectors.toList());
 	}
 	
-	public int getNumberOfComputer() {
+	public long getNumberOfComputer() {
 		return service.getNumberOfComputer();
 	}
