+++ CDB/webapp/src/main/java/com/excilys/cdb/webapp/controller/WebController.java
@@ -90,7 +90,7 @@ public class WebController {
 				.collect(Collectors.toList());
 	}
 	
-	public int getNumberOfComputer() {
+	public long getNumberOfComputer() {
 		return service.getNumberOfComputer();
 	}
